requirements.txt
streamlit
openai
pymupdf
sentence-transformers
scikit-learn
pandas
python-dotenv
google-generativeai
pyresparser